requirements.txt
selenium==3.141.0
beautifulsoup4==4.9.3
python-dotenv==0.17.0
